backend/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncio==3.4.3
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
click==8.1.8
deprecated==1.2.18
distlib==0.3.9
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.17.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
identify==2.6.8
idna==3.10
iniconfig==2.0.0
isort==6.0.0
jinja2==3.1.6
limits==4.0.1
loguru==0.7.3
mako==1.3.9
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
nodeenv==1.9.1
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==4.1.0
psycopg==3.2.5
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.0
pygments==2.19.1
pyjwt==2.10.1
pytest==8.3.4
python-dotenv==1.0.1
python-multipart==0.0.20
pyyaml==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.38
sqlmodel==0.0.23
starlette==0.45.3
trio==0.29.0
typer==0.15.1
typing-extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.2
watchfiles==1.0.4
websockets==15.0
wrapt==1.17.2
